requirements.txt
Flask
Jinja2
langchain
langchain-community
langchain-core
langchain-openai
langchainhub
openai
python-dotenv 